+++ docs/en_US/changelog.md
@@ -4,6 +4,11 @@
 >
 >As a reminder if there is no information on the update, it means that it only concerns the updating of documentation, translation or text.
 
+# 04/03/2022
+- Logs typo fix
+- Useless sudo rights removal
+- Manual refresh in cron to prevent UI lock
+
 # Stable - 18/02/2022
 - Fix on last prints computation